+++ modules/textadept/menu.lua
@@ -97,7 +97,7 @@ local default_menubar = {
     {_L['Complete Word'], function()
       textadept.editing.autocomplete('word')
     end},
-    {_L['Toggle Block Comment'], textadept.editing.block_comment},
+    {_L['Toggle Block Comment'], textadept.editing.toggle_comment},
     {_L['Transpose Characters'], textadept.editing.transpose_chars},
     {_L['Join Lines'], textadept.editing.join_lines},
     {_L['Filter Through'], function()